refactor(moderator dashboard): refactor page components into own folder for more clarity
Each component specific to a page will live inside a directory with the same name as the page
component along with the styles file. The page styles will live in the root styles directory with
the name [page-component-name].module.css.
