+++ components/moderator/OrgDetail/OrgDetail.module.css
@@ -1,0 +1,30 @@
+.root {
+  background-color: white;
+  width: 100%;
+}
+
+.tab {
+  margin-left: 10px;
+}
+
+.row {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: 10px;
+  margin-bottom: 25px;
+}
+
+.response {
+  font-size: 20px;
+  font-weight: bold;
+}
+
+.infoBox {
+  border: 2px solid grey;
+  border-radius: 8px;
+  padding: 6px;
+  margin: 2px;
+  font-size: 15px;
+  font-weight: 100;
+}
