+++ pages/moderator/index.tsx
@@ -1,14 +1,14 @@
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-/* eslint-disable jsx-a11y/no-static-element-interactions */
-import { OrgApp } from 'interfaces';
-import { GetStaticProps } from 'next';
-import { sampleOrgAppData } from 'utils/sample-data';
-import Layout from 'components/Layout';
-import styles from 'styles/index.module.css';
-import OrgCard from 'components/OrgCard';
-import OrgDetail from 'components/OrgDetail';
 import { useState } from 'react';
-import Button from '@material-ui/core/Button';
+import { GetStaticProps } from 'next';
+import { OrgApp } from 'interfaces';
+import Layout from 'components/Layout';
+import OrgCard from 'components/moderator/OrgCard';
+import OrgDetail from 'components/moderator/OrgDetail';
+import { Button } from '@material-ui/core';
+
+import { sampleOrgAppData } from 'utils/sample-data';
+
+import styles from '../../styles/Moderator.module.css';
 
 type Props = {
   items: OrgApp[];
@@ -27,6 +27,8 @@ const ModeratorDashBoard: React.FunctionComponent<Props> = ({ items }) => {
       <div className={styles.root}>
         <div className={styles.leftCol}>
           {items.map((item) => (
+            // TODO: Add accessibility support
+            // eslint-disable-next-line jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events
             <div key={item.id} onClick={() => clickCard(item)}>
               <OrgCard items={item} />
             </div>
